+++ lib/rubocop/cop/thread_safety/dir_chdir.rb
@@ -4,6 +4,8 @@ module RuboCop
   module Cop
     module ThreadSafety
       # Avoid using `Dir.chdir` due to its process-wide effect.
+      # If `AllowCallWithBlock` (disabled by default) option is enabled,
+      # calling `Dir.chdir` with block will be allowed.
       #
       # @example
       #   # bad
@@ -11,6 +13,19 @@ module RuboCop
       #
       #   # bad
       #   FileUtils.chdir("/var/run")
+      #
+      # @example AllowCallWithBlock: false (default)
+      #   # good
+      #   Dir.chdir("/var/run") do
+      #     puts Dir.pwd
+      #   end
+      #
+      # @example AllowCallWithBlock: true
+      #   # bad
+      #   Dir.chdir("/var/run") do
+      #     puts Dir.pwd
+      #   end
+      #
       class DirChdir < Base
         MESSAGE = 'Avoid using `%<module>s.%<method>s` due to its process-wide effect.'
         RESTRICT_ON_SEND = %i[chdir cd].freeze
@@ -24,12 +39,19 @@ module RuboCop
         MATCHER
 
         def on_send(node)
-          chdir?(node) do
-            add_offense(
-              node,
-              message: format(MESSAGE, module: node.receiver.short_name, method: node.method_name)
-            )
-          end
+          return unless chdir?(node)
+          return if allow_call_with_block? && (node.block_argument? || node.parent&.block_type?)
+
+          add_offense(
+            node,
+            message: format(MESSAGE, module: node.receiver.short_name, method: node.method_name)
+          )
+        end
+
+        private
+
+        def allow_call_with_block?
+          !!cop_config['AllowCallWithBlock']
         end
       end
     end